Remove warnings.

+++ SWHex/NSNumber+SWHex.m
@@ -18,7 +18,7 @@
     long result = 0;
     
     int round = 0;
-    int lastDecNum = 0;
+    long lastDecNum = 0;
     while (YES) {
         if (round>0) {
             hexFac *= 16;
